# order-service/auth-service/api-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "E-Commerce API Assignment",
    "version": "1.0",
    "description": "API Documentation for E-Commerce App — Auth & Security Service"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Environment"
    }
  ],
  "security": [
    { "bearerAuth": [] },
    { "cookieAuth": [] }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT via Authorization header (Bearer token)"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token",
        "description": "JWT via access_token cookie"
      }
    },
    "schemas": {
      "GenderEnum": {
        "type": "string",
        "enum": ["MALE", "FEMALE"]
      },
      "StatusEnum": {
        "type": "string",
        "enum": ["ACTIVE", "INACTIVE", "DELETED"]
      },
      "ApiResponseVoid": {
        "type": "object",
        "properties": {
          "code":      { "type": "integer", "example": 1000 },
          "message":   { "type": "string" },
          "result":    { "nullable": true },
          "timestamp": { "type": "string", "format": "date-time" },
          "path":      { "type": "string" }
        }
      },
      "LoginRequestDTO": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email":    { "type": "string", "format": "email", "example": "user@example.com" },
          "password": { "type": "string", "example": "P@ssw0rd!" }
        }
      },
      "LoginResponseDTO": {
        "type": "object",
        "properties": {
          "accessToken":  { "type": "string" },
          "refreshToken": { "type": "string" },
          "expiresIn":    { "type": "integer", "format": "int64", "example": 900000 }
        }
      },
      "ApiResponseLogin": {
        "type": "object",
        "properties": {
          "code":      { "type": "integer", "example": 1000 },
          "message":   { "type": "string" },
          "result":    { "$ref": "#/components/schemas/LoginResponseDTO" },
          "timestamp": { "type": "string", "format": "date-time" },
          "path":      { "type": "string" }
        }
      },
      "ChangePasswordRequestDTO": {
        "type": "object",
        "required": ["oldPassword", "newPassword", "confirmPassword"],
        "properties": {
          "oldPassword":     { "type": "string" },
          "newPassword":     { "type": "string" },
          "confirmPassword": { "type": "string" }
        }
      },
      "RegisterRequestDTO": {
        "type": "object",
        "properties": {
          "email":       { "type": "string", "format": "email" },
          "phoneNumber": { "type": "string" },
          "password":    { "type": "string" },
          "firstName":   { "type": "string" },
          "lastName":    { "type": "string" },
          "gender":      { "$ref": "#/components/schemas/GenderEnum" },
          "birthDate":   { "type": "string", "format": "date" }
        }
      },
      "AdminCreateAccountRequestDTO": {
        "type": "object",
        "properties": {
          "email":       { "type": "string", "format": "email" },
          "phoneNumber": { "type": "string" },
          "password":    { "type": "string" },
          "firstName":   { "type": "string" },
          "lastName":    { "type": "string" },
          "gender":      { "$ref": "#/components/schemas/GenderEnum" },
          "birthDate":   { "type": "string", "format": "date" },
          "roleIds": {
            "type": "array",
            "items": { "type": "string", "format": "uuid" }
          }
        }
      },
      "VerifyOTPRequestDTO": {
        "type": "object",
        "required": ["email", "otp"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "otp":   { "type": "string" }
        }
      },
      "AccountResponseForUserDTO": {
        "type": "object",
        "properties": {
          "id":          { "type": "string", "format": "uuid" },
          "email":       { "type": "string" },
          "phoneNumber": { "type": "string" },
          "status":      { "$ref": "#/components/schemas/StatusEnum" },
          "createdAt":   { "type": "string", "format": "date-time" },
          "updatedAt":   { "type": "string", "format": "date-time" },
          "firstName":   { "type": "string" },
          "lastName":    { "type": "string" },
          "gender":      { "$ref": "#/components/schemas/GenderEnum" },
          "birthDate":   { "type": "string", "format": "date" }
        }
      },
      "PermissionResponseDTO": {
        "type": "object",
        "properties": {
          "id":     { "type": "string", "format": "uuid" },
          "module": { "type": "string" },
          "code":   { "type": "string" }
        }
      },
      "RoleResponseDTO": {
        "type": "object",
        "properties": {
          "id":          { "type": "string", "format": "uuid" },
          "code":        { "type": "string" },
          "name":        { "type": "string" },
          "description": { "type": "string" },
          "permissions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PermissionResponseDTO" }
          }
        }
      },
      "AccountResponseForAdminDTO": {
        "type": "object",
        "properties": {
          "id":          { "type": "string", "format": "uuid" },
          "email":       { "type": "string" },
          "phoneNumber": { "type": "string" },
          "status":      { "$ref": "#/components/schemas/StatusEnum" },
          "createdAt":   { "type": "string", "format": "date-time" },
          "updatedAt":   { "type": "string", "format": "date-time" },
          "firstName":   { "type": "string" },
          "lastName":    { "type": "string" },
          "gender":      { "$ref": "#/components/schemas/GenderEnum" },
          "birthDate":   { "type": "string", "format": "date" },
          "roles": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/RoleResponseDTO" }
          }
        }
      },
      "ApiResponseAccountUser": {
        "type": "object",
        "properties": {
          "code":      { "type": "integer" },
          "message":   { "type": "string" },
          "result":    { "$ref": "#/components/schemas/AccountResponseForUserDTO" },
          "timestamp": { "type": "string", "format": "date-time" },
          "path":      { "type": "string" }
        }
      },
      "ApiResponseAccountAdmin": {
        "type": "object",
        "properties": {
          "code":      { "type": "integer" },
          "message":   { "type": "string" },
          "result":    { "$ref": "#/components/schemas/AccountResponseForAdminDTO" },
          "timestamp": { "type": "string", "format": "date-time" },
          "path":      { "type": "string" }
        }
      },
      "PageAccountResponseForAdminDTO": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/AccountResponseForAdminDTO" }
          },
          "totalElements": { "type": "integer" },
          "totalPages":    { "type": "integer" },
          "size":          { "type": "integer" },
          "number":        { "type": "integer" }
        }
      },
      "ApiResponsePageAccountAdmin": {
        "type": "object",
        "properties": {
          "code":      { "type": "integer" },
          "message":   { "type": "string" },
          "result":    { "$ref": "#/components/schemas/PageAccountResponseForAdminDTO" },
          "timestamp": { "type": "string", "format": "date-time" },
          "path":      { "type": "string" }
        }
      },
...
    }
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login",
        "description": "Authenticate user. Sets `access_token` and `refresh_token` cookies.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequestDTO" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseLogin" }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/change-password": {
      "put": {
        "tags": ["Auth"],
        "summary": "Change password",
        "description": "Changes the current user's password. Clears auth cookies on success.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChangePasswordRequestDTO" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseVoid" }
              }
            }
          }
        }
      }
    },
...
    "/api/accounts/admin": {
      "post": {
        "tags": ["Accounts"],
        "summary": "Create account (Admin)",
        "description": "Requires role `ADMIN` or permission `account:write`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AdminCreateAccountRequestDTO" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseAccountAdmin" }
              }
            }
          }
        }
      }
    },
    "/api/accounts/admin/{id}": {
      "put": {
        "tags": ["Accounts"],
        "summary": "Update account (Admin)",
        "description": "Requires role `ADMIN` or permission `account:write`.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AdminCreateAccountRequestDTO" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseAccountAdmin" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Accounts"],
        "summary": "Get account by ID (Admin)",
        "description": "Requires role `ADMIN` or permission `account:read`.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Account fetched",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseAccountAdmin" }
              }
            }
          }
        }
      }
    },
    "/api/accounts/user/register": {
      "post": {
        "tags": ["Accounts"],
        "summary": "Register user (step 1)",
        "description": "Initiates registration — sends OTP to the provided email.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequestDTO" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Registration initiated, OTP sent",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseVoid" }
              }
            }
          }
        }
      }
    },
    "/api/accounts/user/register/verify-otp": {
      "post": {
        "tags": ["Accounts"],
        "summary": "Verify OTP and complete registration (step 2)",
        "description": "Verifies the OTP and creates the account.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/VerifyOTPRequestDTO" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP verified, account created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseAccountUser" }
              }
            }
          }
        }
      }
    },
    "/api/accounts/user/register/resend-otp": {
      "post": {
        "tags": ["Accounts"],
        "summary": "Resend OTP",
        "description": "Resends the registration OTP to the given email.",
        "security": [],
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "email" }
          }
        ],
        "responses": {
          "200": {
            "description": "OTP resent",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseVoid" }
              }
            }
          }
        }
      }
    },
    "/api/accounts/user/{id}": {
      "put": {
        "tags": ["Accounts"],
        "summary": "Update own account (User)",
        "description": "Requires role `USER`.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequestDTO" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseAccountUser" }
              }
            }
          }
        }
      }
    },
    "/api/accounts/{id}": {
      "delete": {
        "tags": ["Accounts"],
        "summary": "Delete account",
        "description": "Requires role `ADMIN` or permission `account:delete`.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Account deleted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseVoid" }
              }
            }
          }
        }
      }
    },
    "/api/accounts/all": {
      "get": {
        "tags": ["Accounts"],
        "summary": "Get all accounts (paginated)",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 0 }
          },
          {
            "name": "size",
            "in": "query",
            "schema": { "type": "integer", "default": 20 }
          }
        ],
        "responses": {
          "200": {
            "description": "Accounts fetched",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponsePageAccountAdmin" }
              }
            }
          }
        }
      }
    },
...
  }
}